+++ .eslintrc.js
@@ -285,7 +285,12 @@ module.exports = {
   },
   overrides: [
     {
-      files: ["*.stories.tsx", "*.spec.tsx", "*.docsExample.tsx"],
+      files: [
+        "storybook/*",
+        "*.stories.tsx",
+        "*.spec.tsx",
+        "*.docsExample.tsx",
+      ],
       rules: {
         "import/no-extraneous-dependencies": "off",
       },