+++ draft-packages/hero-card/docs/HeroCard.stories.tsx
@@ -1,10 +1,7 @@
 import React from "react"
-import { Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { StoryFn } from "@storybook/react"
 import { Button } from "@kaizen/button"
 import { HeroCard } from "@kaizen/draft-hero-card"
-import { CATEGORIES } from "../../../storybook/constants"
-import { figmaEmbed } from "../../../storybook/helpers"
 
 const ILLUSTRATION_SURVEY = require("./survey.png")
 
@@ -25,7 +22,8 @@ const renderContent = (): JSX.Element => (
 )
 
 export default {
-  title: `${CATEGORIES.deprecated}/Hero Card`,
+  tags: ["autodocs"],
+  title: "Deprecated/Hero Card",
   component: HeroCard,
   parameters: {
     docs: {
@@ -34,25 +32,21 @@ export default {
           '⛔️ This component is deprecated. No further changes will be made to it as it will be superseded by `Tile`.<br/><br/>`import { HeroCard } from "@kaizen/draft-hero-card"`',
       },
     },
-    ...figmaEmbed(
-      "https://www.figma.com/file/eZKEE5kXbEMY3lx84oz8iN/%E2%9D%A4%EF%B8%8F-UI-Kit%3A-Heart?node-id=3568%3A150"
-    ),
   },
-  decorators: [withDesign],
 }
 
-export const DefaultKaizenSiteDemo: Story = () => (
+export const DefaultKaizenSiteDemo: StoryFn = () => (
   <HeroCard>{renderContent()}</HeroCard>
 )
 DefaultKaizenSiteDemo.storyName = "Default (Kaizen Site Demo)"
 
-export const Title: Story = () => (
+export const Title: StoryFn = () => (
   <HeroCard title={<h1>Preview the survey questions</h1>}>
     {renderContent()}
   </HeroCard>
 )
 
-export const Badge: Story = () => (
+export const Badge: StoryFn = () => (
   <HeroCard
     title={<h1>Preview the survey questions</h1>}
     badge={<span>1</span>}
@@ -62,7 +56,7 @@ export const Badge: Story = () => (
 )
 Badge.parameters = { chromatic: { disable: false } }
 
-export const Image: Story = () => (
+export const Image: StoryFn = () => (
   <HeroCard
     title={<h1>Preview the survey questions</h1>}
     badge={<span>1</span>}
@@ -84,7 +78,7 @@ export const Image: Story = () => (
 )
 Image.parameters = { chromatic: { disable: false } }
 
-export const CustomLeftContent: Story = () => (
+export const CustomLeftContent: StoryFn = () => (
   <HeroCard
     title={<h1>Preview the survey questions</h1>}
     leftContent={<p>Ta-dah</p>}
@@ -94,7 +88,7 @@ export const CustomLeftContent: Story = () => (
 )
 CustomLeftContent.parameters = { chromatic: { disable: false } }
 
-export const CustomLeftContentAndBadge: Story = () => (
+export const CustomLeftContentAndBadge: StoryFn = () => (
   <HeroCard
     title={<h1>Preview the survey questions</h1>}
     leftContent={<p>Ta-dah</p>}
@@ -105,7 +99,7 @@ export const CustomLeftContentAndBadge: Story = () => (
 )
 CustomLeftContentAndBadge.parameters = { chromatic: { disable: false } }
 
-export const FullWidth: Story = () => (
+export const FullWidth: StoryFn = () => (
   <HeroCard
     title={<h1>Preview the survey questions</h1>}
     leftContent={<p>Ta-dah</p>}
@@ -116,7 +110,7 @@ export const FullWidth: Story = () => (
 )
 FullWidth.parameters = { chromatic: { disable: false } }
 
-export const BackgroundColors: Story = () => (
+export const BackgroundColors: StoryFn = () => (
   <HeroCard
     title={<h1>Preview the survey questions</h1>}
     leftBackgroundColor="cluny200"