+++ draft-packages/tag/KaizenDraft/Tag/index.ts
@@ -1,1 +1,1 @@
-export { default as Tag } from "./Tag"
+export * from "./Tag"