+++ draft-packages/title-block-zen/docs/TitleBlockZen.stories.tsx
@@ -1,6 +1,5 @@
 import React, { ReactNode } from "react"
-import { ComponentStory, Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { Meta, StoryFn } from "@storybook/react"
 import { Box } from "@kaizen/component-library"
 import addIcon from "@kaizen/component-library/icons/add.icon.svg"
 import arrowForwardIcon from "@kaizen/component-library/icons/arrow-forward.icon.svg"
@@ -11,15 +10,15 @@ import { Container, Content, Skirt, SkirtCard } from "@kaizen/draft-page-layout"
 import { Tag } from "@kaizen/draft-tag"
 import { assetUrl } from "@kaizen/hosted-assets"
 import { Heading, Paragraph } from "@kaizen/typography"
-import { CATEGORIES } from "../../../storybook/constants"
-import { figmaEmbed } from "../../../storybook/helpers"
 import { NavigationTab, SectionTitleRenderProps, TitleBlockZen } from ".."
 import styles from "./TitleBlockZen.stories.module.scss"
 
 const TESTING_VIEWPORTS = [320, 768, 1200]
 
 export default {
-  title: `${CATEGORIES.components}/Title Block`,
+  tags: ["autodocs"],
+  title: "Components/Title Block",
+  component: TitleBlockZen,
   parameters: {
     chromatic: { viewports: TESTING_VIEWPORTS },
     docs: {
@@ -28,12 +27,8 @@ export default {
           'import { TitleBlockZen } from "@kaizen/draft-title-block-zen"',
       },
     },
-    ...figmaEmbed(
-      "https://www.figma.com/file/GMxm8rvDCbj0Xw3TQWBZ8b/UI-Kit-Zen?node-id=4619%3A17068"
-    ),
   },
-  decorators: [withDesign],
-}
+} satisfies Meta<typeof TitleBlockZen>
 
 const OffsetPadding = ({
   children,
@@ -62,14 +57,13 @@ const SECONDARY_ACTIONS = [
   },
 ]
 
-const DefaultTemplate: ComponentStory<typeof TitleBlockZen> = args => (
+const DefaultTemplate: StoryFn<typeof TitleBlockZen> = args => (
   <OffsetPadding>
     <TitleBlockZen {...args} />
   </OffsetPadding>
 )
 
-export const Default: ComponentStory<typeof TitleBlockZen> =
-  DefaultTemplate.bind({})
+export const Default: StoryFn<typeof TitleBlockZen> = DefaultTemplate.bind({})
 Default.args = {
   title: "Page title",
   surveyStatus: { text: "Live", status: "live" },
@@ -112,7 +106,7 @@ Default.args = {
   ],
 }
 
-export const WithBadge: Story = () => {
+export const WithBadge: StoryFn = () => {
   const [badgeCount, setBadgeCount] = React.useState(1)
   return (
     <OffsetPadding>
@@ -147,7 +141,7 @@ export const WithBadge: Story = () => {
 WithBadge.storyName = "With Primary Action Badge"
 WithBadge.parameters = { chromatic: { disable: false } }
 
-export const WithDefaultTag: Story = () => (
+export const WithDefaultTag: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -169,7 +163,7 @@ export const WithDefaultTag: Story = () => (
 WithDefaultTag.storyName = "With Default Survey Status (Tag)"
 WithDefaultTag.parameters = { chromatic: { disable: false } }
 
-export const AdminWithDefaultTag: Story = () => (
+export const AdminWithDefaultTag: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       variant="admin"
@@ -192,7 +186,7 @@ export const AdminWithDefaultTag: Story = () => (
 AdminWithDefaultTag.storyName = "Admin With Default Survey Status (Tag)"
 AdminWithDefaultTag.parameters = { chromatic: { disable: false } }
 
-export const DefaultWithMenuButton: Story = () => (
+export const DefaultWithMenuButton: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -239,7 +233,7 @@ export const DefaultWithMenuButton: Story = () => (
 DefaultWithMenuButton.storyName = "Default (Menu Button)"
 DefaultWithMenuButton.parameters = { chromatic: { disable: false } }
 
-export const AdminVariant: Story = () => (
+export const AdminVariant: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -273,7 +267,7 @@ export const AdminVariant: Story = () => (
 AdminVariant.storyName = "Admin variant"
 AdminVariant.parameters = { chromatic: { disable: false } }
 
-export const AdminVariantWithNavTabs: Story = () => (
+export const AdminVariantWithNavTabs: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -301,7 +295,7 @@ export const AdminVariantWithNavTabs: Story = () => (
 AdminVariantWithNavTabs.storyName = "Admin variant with Navigation Tabs"
 AdminVariantWithNavTabs.parameters = { chromatic: { disable: false } }
 
-export const EducationVariant: Story = () => (
+export const EducationVariant: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -357,7 +351,7 @@ export const EducationVariant: Story = () => (
 EducationVariant.storyName = "Education variant"
 EducationVariant.parameters = { chromatic: { disable: false } }
 
-export const Engagement: Story = () => (
+export const Engagement: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Baseline Engagement Survey"
@@ -396,7 +390,7 @@ export const Engagement: Story = () => (
 )
 Engagement.parameters = { chromatic: { disable: false } }
 
-export const Performance: Story = () => (
+export const Performance: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Blanca Wheeler"
@@ -444,7 +438,7 @@ export const Performance: Story = () => (
 )
 Performance.parameters = { chromatic: { disable: false } }
 
-export const PerformanceWithAvatarProps: Story = () => (
+export const PerformanceWithAvatarProps: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Blanca Wheeler"
@@ -495,7 +489,7 @@ export const PerformanceWithAvatarProps: Story = () => (
 )
 PerformanceWithAvatarProps.storyName = "Performance with AvatarProps"
 
-export const PerformanceWithEmptyAvatarProps: Story = () => (
+export const PerformanceWithEmptyAvatarProps: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Blanca Wheeler"
@@ -543,7 +537,7 @@ export const PerformanceWithEmptyAvatarProps: Story = () => (
 )
 PerformanceWithEmptyAvatarProps.storyName = "Performance with Empty AvatarProps"
 
-export const LongLabels: Story = () => (
+export const LongLabels: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Wolfeschlegelsteino Hausenbergerdorffsch Hausenbergerdorffsch"
@@ -614,7 +608,7 @@ const MENU_LINKS = [
   },
 ]
 
-export const DefaultWithContent: Story = () => (
+export const DefaultWithContent: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -728,7 +722,7 @@ export const DefaultWithContent: Story = () => (
 DefaultWithContent.storyName = "Default with content"
 DefaultWithContent.parameters = { chromatic: { disable: false } }
 
-export const DefaultNoSecondary: Story = () => (
+export const DefaultNoSecondary: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -839,7 +833,7 @@ export const DefaultNoSecondary: Story = () => (
 )
 DefaultNoSecondary.storyName = "Default (no secondary actions)"
 
-export const DefaultOnlyPrimary: Story = () => (
+export const DefaultOnlyPrimary: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -868,7 +862,7 @@ export const DefaultOnlyPrimary: Story = () => (
 )
 DefaultOnlyPrimary.storyName = "Default (only primary action)"
 
-export const DefaultOnlySecondary: Story = () => (
+export const DefaultOnlySecondary: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -892,7 +886,7 @@ export const DefaultOnlySecondary: Story = () => (
   </OffsetPadding>
 )
 DefaultOnlySecondary.storyName = "Default (only secondary action)"
-export const DefaultWithReportSwitcher: Story = () => (
+export const DefaultWithReportSwitcher: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
@@ -945,7 +939,7 @@ export const DefaultWithReportSwitcher: Story = () => (
 DefaultWithReportSwitcher.storyName = "Default with report switcher"
 DefaultWithReportSwitcher.parameters = { chromatic: { disable: false } }
 
-export const DefaultNoLink: Story = () => (
+export const DefaultNoLink: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -1000,7 +994,7 @@ export const DefaultNoLink: Story = () => (
 )
 DefaultNoLink.storyName = "Default (no link in breadcrumb)"
 
-export const DefaultOnlyLongTitle: Story = () => (
+export const DefaultOnlyLongTitle: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
@@ -1016,7 +1010,7 @@ export const DefaultOnlyLongTitle: Story = () => (
 )
 DefaultOnlyLongTitle.storyName = "Default (only long title)"
 
-export const DefaultCollapsedNavigation: Story = () => (
+export const DefaultCollapsedNavigation: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -1041,7 +1035,7 @@ export const DefaultCollapsedNavigation: Story = () => (
 DefaultCollapsedNavigation.storyName = "Default (collapsed navigation)"
 DefaultCollapsedNavigation.parameters = { chromatic: { disable: false } }
 
-export const DefaultCollapsedNavigationCard: Story = () => (
+export const DefaultCollapsedNavigationCard: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -1068,7 +1062,7 @@ export const DefaultCollapsedNavigationCard: Story = () => (
 DefaultCollapsedNavigationCard.storyName =
   "Default (collapsed navigation with card)"
 
-export const AdminVariantNavigation: Story = () => (
+export const AdminVariantNavigation: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       variant="admin"
@@ -1095,7 +1089,7 @@ export const AdminVariantNavigation: Story = () => (
 )
 AdminVariantNavigation.storyName = "Admin (collapsed navigation)"
 
-export const RenderProps: Story = () => {
+export const RenderProps: StoryFn = () => {
   const CustomTab = (props: {
     href: string
     className: string
@@ -1147,7 +1141,7 @@ export const RenderProps: Story = () => {
   )
 }
 
-export const WithCustomSectionTitle: Story = () => {
+export const WithCustomSectionTitle: StoryFn = () => {
   const CustomComponent = (props: SectionTitleRenderProps): JSX.Element => (
     <div className={styles["flex-wrapper"]}>
       <Heading color="white" variant="heading-3">
@@ -1214,7 +1208,7 @@ const MockRouterLink = ({
   </button>
 )
 
-export const ActionRenderProps: Story = () => (
+export const ActionRenderProps: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"
@@ -1256,7 +1250,7 @@ export const ActionRenderProps: Story = () => (
 
 ActionRenderProps.storyName = "Custom actions with component render props"
 
-export const MenuHierarchyExample: Story = () => (
+export const MenuHierarchyExample: StoryFn = () => (
   <OffsetPadding>
     <TitleBlockZen
       title="Page title"