+++ packages/date-picker/docs/filter-button-foundations/FilterButtonGroup.stories.tsx
@@ -1,17 +1,13 @@
 import React from "react"
-import { ComponentMeta, ComponentStory, Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { Meta, StoryFn } from "@storybook/react"
 import { StickerSheet } from "../../../../storybook/components/StickerSheet"
-import {
-  CATEGORIES,
-  SUB_CATEGORIES,
-  SUB_COMPONENTS_FOLDER_NAME,
-} from "../../../../storybook/constants"
 import { FilterBaseButton } from "../../src/FilterDateRangePicker/components/Trigger/FilterBaseButton"
 import { FilterButtonGroup } from "../../src/FilterDateRangePicker/components/Trigger/FilterButtonGroup"
 
 export default {
-  title: `${CATEGORIES.components}/${SUB_CATEGORIES.datePicker}/Filter Date Range Picker/${SUB_COMPONENTS_FOLDER_NAME}/Filter Button Group`,
+  tags: ["autodocs"],
+  title:
+    "Components/Filter Date Range Picker/Subcomponents/Filter Button Group",
   component: FilterButtonGroup,
   parameters: {
     docs: {
@@ -20,10 +16,9 @@ export default {
       },
     },
   },
-  decorators: [withDesign],
-} as ComponentMeta<typeof FilterButtonGroup>
+} satisfies Meta<typeof FilterButtonGroup>
 
-export const DefaultStory: ComponentStory<typeof FilterButtonGroup> = args => (
+export const DefaultStory: StoryFn<typeof FilterButtonGroup> = args => (
   <FilterButtonGroup {...args}>
     <FilterBaseButton>First</FilterBaseButton>
     <FilterBaseButton>Last</FilterBaseButton>
@@ -31,7 +26,7 @@ export const DefaultStory: ComponentStory<typeof FilterButtonGroup> = args => (
 )
 DefaultStory.storyName = "Filter Button Group"
 
-const StickerSheetTemplate: Story = () => (
+const StickerSheetTemplate: StoryFn = () => (
   <>
     <StickerSheet heading="Filter Button Group">
       <StickerSheet.Header headings={["Group of 2", "Group of 3"]} />