+++ packages/date-picker/docs/filter-trigger-buttons/filter-trigger-buttons.stories.tsx
@@ -1,23 +1,15 @@
 import React from "react"
-import { Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { Meta, StoryFn } from "@storybook/react"
 import { StickerSheet } from "../../../../storybook/components/StickerSheet"
-import {
-  CATEGORIES,
-  SUB_CATEGORIES,
-  SUB_COMPONENTS_FOLDER_NAME,
-} from "../../../../storybook/constants"
 import {
   FilterTriggerButton,
   RemovableFilterTriggerButton,
 } from "../../src/FilterDateRangePicker/components/Trigger"
 
 export default {
-  title: `${CATEGORIES.components}/${SUB_CATEGORIES.datePicker}/Filter Date Range Picker/${SUB_COMPONENTS_FOLDER_NAME}/Filter Trigger Buttons`,
-  subcomponents: {
-    FilterTriggerButton,
-    RemovableFilterTriggerButton,
-  },
+  tags: ["autodocs"],
+  title:
+    "Components/Filter Date Range Picker/Subcomponents/Filter Trigger Buttons",
   parameters: {
     docs: {
       description: {
@@ -25,10 +17,9 @@ export default {
       },
     },
   },
-  decorators: [withDesign],
-}
+} satisfies Meta
 
-const StickerSheetTemplate: Story = () => (
+const StickerSheetTemplate: StoryFn = () => (
   <>
     <StickerSheet heading="Filter Trigger Button">
       <StickerSheet.Header