+++ packages/notification/docs/InlineNotification.stories.tsx
@@ -1,14 +1,12 @@
 import React from "react"
-import { ComponentStory, Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { StoryFn } from "@storybook/react"
 import { InlineNotification } from "@kaizen/notification"
 import { Heading, HeadingProps } from "@kaizen/typography"
 import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
-import { CATEGORIES, SUB_CATEGORIES } from "../../../storybook/constants"
-import { figmaEmbed } from "../../../storybook/helpers"
 
 export default {
-  title: `${CATEGORIES.components}/${SUB_CATEGORIES.notification}/Inline Notification`,
+  tags: ["autodocs"],
+  title: "Components/Notification/Inline Notification",
   component: InlineNotification,
   parameters: {
     docs: {
@@ -16,16 +14,10 @@ export default {
         component: 'import { InlineNotification } from "@kaizen/notification";',
       },
     },
-    ...figmaEmbed(
-      "https://www.figma.com/file/eZKEE5kXbEMY3lx84oz8iN/%E2%9D%A4%EF%B8%8F-UI-Kit%3A-Heart?node-id=13877%3A66008"
-    ),
   },
-  decorators: [withDesign],
 }
 
-export const DefaultKaizenDemo: ComponentStory<
-  typeof InlineNotification
-> = props => (
+export const DefaultKaizenDemo: StoryFn<typeof InlineNotification> = props => (
   <InlineNotification {...props}>
     New user data, imported by mackenzie@hooli.com has successfully uploaded.{" "}
     <a href="/">Manage users is now available</a>
@@ -42,9 +34,7 @@ const customHeadingProps: HeadingProps = {
   tag: "h2",
   children: "Custom",
 }
-export const CustomHeadingLevel: ComponentStory<
-  typeof InlineNotification
-> = props => (
+export const CustomHeadingLevel: StoryFn<typeof InlineNotification> = props => (
   <InlineNotification headingProps={customHeadingProps} {...props}>
     New user data
   </InlineNotification>
@@ -54,7 +44,7 @@ CustomHeadingLevel.args = {
   type: "positive",
 }
 
-const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
+const StickerSheetTemplate: StoryFn<{ isReversed: boolean }> = ({
   isReversed,
 }) => {
   const forceMultilineText = (
@@ -95,6 +85,17 @@ const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
             "All Employees - North America" status has been changed to
             'Archived'.
             <a href="/">View all</a>
+          </InlineNotification>
+        </StoryWrapper.Row>
+        <StoryWrapper.Row rowTitle="Subtle">
+          <InlineNotification
+            type="informative"
+            isSubtle
+            title="Informative title"
+            hideCloseIcon
+          >
+            "All Employees - North America" status has been changed to
+            'Archived'. <a href="/">View all</a>
           </InlineNotification>
         </StoryWrapper.Row>
         <StoryWrapper.Row rowTitle="Positive">
@@ -164,9 +165,7 @@ StickerSheetReversed.parameters = {
   controls: { disable: true },
 }
 
-export const AutohideDemo: ComponentStory<
-  typeof InlineNotification
-> = props => (
+export const AutohideDemo: StoryFn<typeof InlineNotification> = props => (
   <>
     <InlineNotification
       title="Success"