+++ packages/rich-text-editor/docs/Toolbar.stories.tsx
@@ -1,22 +1,16 @@
 import React from "react"
-import { ComponentStory } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { StoryFn } from "@storybook/react"
 import linkIcon from "@kaizen/component-library/icons/add-link.icon.svg"
 import boldIcon from "@kaizen/component-library/icons/bold.icon.svg"
 import bulletListIcon from "@kaizen/component-library/icons/bulletted-list.icon.svg"
 import italicIcon from "@kaizen/component-library/icons/italics.icon.svg"
 import numberListIcon from "@kaizen/component-library/icons/numbered-list.icon.svg"
 import underlineIcon from "@kaizen/component-library/icons/underline.icon.svg"
-import {
-  CATEGORIES,
-  SUB_CATEGORIES,
-  SUB_COMPONENTS_FOLDER_NAME,
-} from "../../../storybook/constants"
-import { figmaEmbed } from "../../../storybook/helpers"
 import { ToggleIconButton, Toolbar, ToolbarSection } from "../"
 
 export default {
-  title: `${CATEGORIES.components}/${SUB_CATEGORIES.richTextEditor}/${SUB_COMPONENTS_FOLDER_NAME}/Toolbar`,
+  tags: ["autodocs"],
+  title: "Components/Rich Text Editor/Subcomponents/Toolbar",
   component: Toolbar,
   parameters: {
     actions: {
@@ -27,14 +21,10 @@ export default {
         component: 'import { Toolbar } from "@kaizen/rich-text-editor"',
       },
     },
-    ...figmaEmbed(
-      "https://www.figma.com/file/eZKEE5kXbEMY3lx84oz8iN/%E2%9D%A4%EF%B8%8F-UI-Kit%3A-Heart?node-id=12317%3A92573"
-    ),
   },
-  decorators: [withDesign],
 }
 
-const ToolbarStoryTemplate: ComponentStory<typeof Toolbar> = args => (
+const ToolbarStoryTemplate: StoryFn<typeof Toolbar> = args => (
   <Toolbar {...args}>
     <ToolbarSection>
       <ToggleIconButton label="Bold" icon={boldIcon} />