+++ draft-packages/filter-menu-button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.44](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.6.43...@kaizen/draft-filter-menu-button@2.6.44) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
+## [2.6.43](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.6.42...@kaizen/draft-filter-menu-button@2.6.43) (2023-04-05)
+
+**Note:** Version bump only for package @kaizen/draft-filter-menu-button
+
+
+
+
+
 ## [2.6.42](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-menu-button@2.6.41...@kaizen/draft-filter-menu-button@2.6.42) (2023-03-30)
 
 **Note:** Version bump only for package @kaizen/draft-filter-menu-button