+++ draft-packages/title-block-zen/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [7.9.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@7.9.20...@kaizen/draft-title-block-zen@7.9.21) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-title-block-zen
+
+
+
+
+
+## [7.9.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@7.9.19...@kaizen/draft-title-block-zen@7.9.20) (2023-04-05)
+
+**Note:** Version bump only for package @kaizen/draft-title-block-zen
+
+
+
+
+
 ## [7.9.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-title-block-zen@7.9.18...@kaizen/draft-title-block-zen@7.9.19) (2023-03-30)
 
 **Note:** Version bump only for package @kaizen/draft-title-block-zen