+++ packages/a11y/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.7.8...@kaizen/a11y@1.7.9) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
+## [1.7.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.7.7...@kaizen/a11y@1.7.8) (2023-04-05)
+
+**Note:** Version bump only for package @kaizen/a11y
+
+
+
+
+
 ## [1.7.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.7.6...@kaizen/a11y@1.7.7) (2023-03-29)
 
 **Note:** Version bump only for package @kaizen/a11y