+++ packages/button/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/button@2.1.1...@kaizen/button@2.1.2) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/button
+
+
+
+
+
+## [2.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/button@2.1.0...@kaizen/button@2.1.1) (2023-04-05)
+
+**Note:** Version bump only for package @kaizen/button
+
+
+
+
+
 # [2.1.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/button@2.0.29...@kaizen/button@2.1.0) (2023-03-30)