+++ packages/tailwind/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.6.4...@kaizen/tailwind@0.6.5) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
 ## [0.6.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.6.3...@kaizen/tailwind@0.6.4) (2023-03-31)
 
 **Note:** Version bump only for package @kaizen/tailwind